rename: 훅 이름 내 중복의미 삭제
